*DONE*
Added the shared arithmetic module
All the files are just dependencies. you will instance "sharedComplexMUL" and add 10MHz registers after it

the parameters are:
p_inputWidth:      width of real or imaginary data ALONE. ie, a value of 8 will make the input ports 16 for {real,imag} number
p_PointPosition: number of bits from p_inputWidth that are specified for the decimals. The output will have the same number of decimal points as well